move functions to preprocessing
